pyproject.toml
[tool.poetry]
name = "outdoor-sports-planner"
version = "0.1.0"
description = "Outdoor sports planner application that helps to plan sports given weather and airquality"
authors = ["Christian <christianresch@mailbox.org>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.0"
requests = "^2.32.3"
pytest = "^8.3.3"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
openmeteo-requests = "^1.3.0"
requests-cache = "^1.2.1"
pandas = "^2.2.3"
retry-requests = "^2.0.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
pytest-httpx = "^0.35.0"
pytest-asyncio = "^0.25.3"
pika = "^1.3.2"
prometheus-fastapi-instrumentator = "^7.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
